requirements.txt
beautifulsoup4>=4.12.0
ebooklib>=0.17.1
fastapi>=0.100.0
jinja2>=3.1.2
uvicorn>=0.23.0
python-multipart>=0.0.6
pymupdf>=1.23.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
